fix shaders

+++ ShaderLibrary/Surface.hlsl
@@ -4,17 +4,17 @@
 struct CustomLitData
 {
     float3 positionWS;
-    half3  V; //ViewDirWS
-    half3  N; //NormalWS
-    half3  B; //BinormalWS
-    half3  T; //TangentWS
+    float3  V; //ViewDirWS
+    float3  N; //NormalWS
+    float3  B; //BinormalWS
+    float3  T; //TangentWS
     //float2 ScreenUV;
 };
 
 struct CustomSurfaceData
 {
-    half3 albedo;
-    half3 specular;
+    float3 albedo;
+    float3 specular;
    // half3 normalTS;
   //  float2 uvs;
     half  metallic;
